* dbtarantool.ListInterfaceToStruct

+++ pkg/dbtarantool/common.go
@@ -32,7 +32,8 @@ func ListInterfaceToStruct(listInterface *[]interface{}, mapNameColumnToData map
 		case *uint64:
 			*valueTo = (*listInterface)[idColumn].(uint64)
 		case *int:
-			*valueTo = (*listInterface)[idColumn].(int)
+			// приемник с типом int
+			*valueTo = int((*listInterface)[idColumn].(uint64))
 		case *CustomTime: // uint64
 			*valueTo = CustomTime((*listInterface)[idColumn].(uint64))
 		default:
